Renamed index to swears.js
